3/25 홈 인기 주류 수정 / 신고 기능 구현 중

+++ sool/src/main/java/com/example/sool/dto/DrinkSearchDto.java
@@ -10,8 +10,8 @@ public class DrinkSearchDto {
 
     private String categoryCode;
 
-    private Integer abvLow;
-    private Integer abvHigh;
+    private double abvLow;
+    private double abvHigh;
     
     private Integer priceLow;
     private Integer priceHigh;
